refactor: change naming (axios -> requestInstance)
